Refactor Code 2

+++ src/main/java/ru/shirkov/repositore/Repo.java
@@ -1,5 +1,5 @@
 package ru.shirkov.repositore;
 
 public interface Repo<T> {
-T getById(Long id);
+T getById(Integer id);
 }
